requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.1
aiosqlite>=0.19.0
python-dotenv>=1.0.0
